+++ secrets/data-project-2-449815-02b4b3af0560.json
@@ -1,0 +1,13 @@
+{
+  "type": "service_account",
+  "project_id": "data-project-2-449815",
+  "private_key_id": "02b4b3af0560c35c900ff1f884cae01d5819d5f5",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDFl1bb1Zck2vDN\nP6mSLDRiLPWYxItJLaWLOjUXmEzNaDYTAYgQJEha8j3ZQITt5Sj8oEUjxbhk9l5S\ngNv+HqzBqwI+cz9VCZZ1c1MrzOgEOLEy3MbEHPcbpXRmkxtzf4f5nqjRe4rRm3KQ\nWHi8hEe6eL/fq8Wx6tKE2dXvDMlKxo8XXuM+GPKQy0p4s4V8ccheRUmi/CZRE45F\n/i70N4mptaeg4/vmK63YNgsk2zpjgM4zWkL2tVl8iwECP2fr1BvgbH6pA5VGCPHI\nSkKS/VvOKIhVH1S8vKZecfpWPQrPldtoV/lTUOVezjvIaNrEPFBZjEbt+DfhEErC\nWJv1jsjJAgMBAAECggEAE/dW2jQ3N4zNpinbklcX7H8/SoJ60ihUCPF27bVQtnVk\nOG+wTnRCIPmHPSmnPQxJ49w0nFiliUM5dQQLJt615YSIn6Og59CGFQR0KDIGlG2f\nn0R+q9DC+hbSIC7XiGhfxr4s8C0YEhXvgc1E1wTVZC8BMYJ/obq5Lt49v1AHE6B3\nDo+AhSR1jE7sBPLG6A+IsZ3bpih0icW1dbUKTspNf5MTxGZKpFlWcS7ib8eyHc2r\nJiT7UJUOr+cPv3igzZcNYqo2pxVsFTbRu55KawxVG3A7HOwBGddipA7xmbxvYesU\n2aySKOnJoHbxV3KyVeHmBfLKwEP7v8UrJnn9cI6JwQKBgQDtSMivoZovhZT6Htmp\nW4VXjOgvsLX9oHDi/l1kEbF2gMCMBQ0AMlPKpYpsH7plHIXAezf1ptfyUUgEUsPf\nkQbjiLlB1rnAqCC9YQATj4QDTL4+XjOxtqwkW8oh70u0qZD6duTY2H8G22wJ8Dkf\n7YJaTCr+Ff3L2SsywLAxDn8TQQKBgQDVLRODpkaxRUkZHE1/ap7MXGy+BROaTGX8\n+rrhZjuhRaXGkcgql7dnjaL4zwNj6UyhiSQ9woGxmrClVFbhVFZInRDlR/7b3VKQ\n8M9+wRwg4eO6NBwT6Fe+T/dNmQug1FpjCaRrth3XwgKAnKr+2PqR7Clfzbhiyq4f\nPAnsSj27iQKBgDWYkhPxqZUAVEUmwL1bOcQrxKo4/PJ70ePD1q5qvK3qqRM3dNdN\nhtWoDJsdk4mRJrjyXBJ1jCNkrqKjWzDdXjP4xveTG8Pjdfmwtka0zPm91N7bgkuq\n2jkgLoj5txzojCG2xzwfspmG6GFaTFzk1eO1fqSDrR7lP7eoBGQgRR3BAoGAPTyH\noIfBQucOSoJQ2Ms99qKOMj+vE+HiYjt8N4dwlbE2H8Z8449Vhg+vwx1kcezwsG5+\n1KbtLqGrILMLigr6Bj3oPpTh5zK6pxwl3tmGRQPesPyLHETf/7Fk7wuJxZd6p3x7\nmN1l0TbQFTXsAO0/fW+DqxujXe2u+xX6G8Bv31ECgYEA3IdkuteLJX1mYpDNeTG9\nxjT/z+/rhfFt+zpvi/SfbL5lOgmbO0kRZ532m1snDnQLbyd8GqlfgwERAMjQKbb0\nyLIooeepnplFQgQTLCtGaTkHsiTkuOK63CbhBk8YPqE9Lf5hNIjgT6PFB1tZ8cRs\nzCitPd2zTeK0hpYGi+Gh03w=\n-----END PRIVATE KEY-----\n",
+  "client_email": "pub-sub-publisher@data-project-2-449815.iam.gserviceaccount.com",
+  "client_id": "114526948251363249584",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/pub-sub-publisher%40data-project-2-449815.iam.gserviceaccount.com",
+  "universe_domain": "googleapis.com"
+}
